backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
git+https://github.com/openai/whisper.git
python-multipart>=0.0.9
pydantic>=2.9.0,<3.0.0
numpy<2.0.0
httpx>=0.27.0
python-dotenv>=1.0.1

